Move internal files to internal package
