+++ refreshVersions/src/main/kotlin/de/fayard/versions/internal/ArtifactGrouping.kt
@@ -1,4 +1,4 @@
-package de.fayard.versions
+package de.fayard.versions.internal
 
 /**
  * We assume each part of the "group" is dot separated (`.`), and each part of the name is dash separated (`-`).
