+++ refreshVersions/src/main/kotlin/de/fayard/versions/internal/UsedRepositoriesWriting.kt
@@ -1,4 +1,4 @@
-package de.fayard.versions
+package de.fayard.versions.internal
 
 import de.fayard.versions.extensions.isBuildSrc
 import de.fayard.versions.extensions.isRootProject
