+++ refreshVersions/src/main/kotlin/de/fayard/versions/internal/Version.kt
@@ -1,3 +1,3 @@
-package de.fayard.versions
+package de.fayard.versions.internal
 
 internal inline class Version(val value: String)
