+++ refreshVersions/src/main/kotlin/de/fayard/versions/internal/VersionsPlaceholdersReplacement.kt
@@ -1,6 +1,6 @@
-package de.fayard.versions
+package de.fayard.versions.internal
 
-import de.fayard.versions.ArtifactGroupNaming.*
+import de.fayard.versions.internal.ArtifactGroupNaming.*
 import de.fayard.versions.extensions.isBuildSrc
 import de.fayard.versions.extensions.isGradlePlugin
 import org.gradle.api.Project
